+++ src/RecursoC.cpp
@@ -7,7 +7,7 @@
 
 #include "RecursoC.hpp"
 
-RecursoC::RecursoC(float x,float y) :Punto(x,y){
+RecursoC::RecursoC(double x,double y) :Punto(x,y){
     bubicado=false;
     dist=10E9;
 }
